tests/krate/publish: Extract `timestamps` module
